+++ examples/resources/xray_operational_risk_policy/resource.tf
@@ -1,17 +1,15 @@
-resource "xray_license_policy" "allowed_licenses" {
-  name        = "test-license-policy-allowed"
-  description = "License policy, allow certain licenses"
-  type        = "license"
+resource "xray_operational_risk_policy" "min_risk" {
+  name        = "test-operational-risk-policy-min-risk"
+  description = "Operational Risk policy with a custom risk rule"
+  type        = "Operational_Risk"
   project_key = "testproj"
 
   rule {
-    name     = "License_rule"
+    name     = "op_risk_custom_rule"
     priority = 1
 
     criteria {
-      allowed_licenses         = ["Apache-1.0", "Apache-2.0"]
-      allow_unknown            = false
-      multi_license_permissive = true
+      op_risk_min_risk = "Medium"
     }
 
     actions {
@@ -22,32 +20,37 @@ resource "xray_license_policy" "allowed_licenses" {
       notify_watch_recipients            = true
       notify_deployer                    = true
       create_ticket_enabled              = false // set to true only if Jira integration is enabled
-      custom_severity                    = "High"
       build_failure_grace_period_in_days = 5 // use only if fail_build is enabled
 
       block_download {
         unscanned = true
         active    = true
       }
-
     }
   }
 }
 
-resource "xray_license_policy" "banned_licenses" {
-  name        = "test-license-policy-banned"
-  description = "License policy, block certain licenses"
-  type        = "license"
+resource "xray_operational_risk_policy" "custom_criteria" {
+  name        = "test-operational-risk-policy-custom-criteria"
+  description = "Operational Risk policy with a custom risk rule"
+  type        = "Operational_Risk"
   project_key = "testproj"
 
   rule {
-    name     = "License_rule"
+    name     = "op_risk_custom_rule"
     priority = 1
 
     criteria {
-      banned_licenses          = ["Apache-3.0", "Apache-4.0"]
-      allow_unknown            = false
-      multi_license_permissive = false
+      op_risk_custom {
+        use_and_condition                  = true
+        is_eol                             = false
+        release_date_greater_than_months   = 6
+        newer_versions_greater_than        = 1
+        release_cadence_per_year_less_than = 1
+        commits_less_than                  = 10
+        committers_less_than               = 1
+        risk                               = "medium"
+      }
     }
 
     actions {
@@ -58,7 +61,6 @@ resource "xray_license_policy" "banned_licenses" {
       notify_watch_recipients            = true
       notify_deployer                    = true
       create_ticket_enabled              = false // set to true only if Jira integration is enabled
-      custom_severity                    = "Medium"
       build_failure_grace_period_in_days = 5 // use only if fail_build is enabled
 
       block_download {