+++ examples/resources/xray_watch/resource.tf
@@ -2,6 +2,7 @@ resource "xray_watch" "all-repos" {
   name        = "all-repos-watch"
   description = "Watch for all repositories, matching the filter"
   active      = true
+  project_key = "testproj"
 
   watch_resource {
     type = "all-repos"
@@ -29,6 +30,7 @@ resource "xray_watch" "repository" {
   name        = "repository-watch"
   description = "Watch a single repo or a list of repositories"
   active      = true
+  project_key = "testproj"
 
   watch_resource {
     type       = "repository"
@@ -71,6 +73,7 @@ resource "xray_watch" "all-builds-with-filters" {
   name        = "build-watch"
   description = "Watch all builds with Ant patterns filter"
   active      = true
+  project_key = "testproj"
 
   watch_resource {
     type       = "all-builds"
@@ -99,6 +102,7 @@ resource "xray_watch" "build" {
   name        = "build-watch"
   description = "Watch a single build or a list of builds"
   active      = true
+  project_key = "testproj"
 
   watch_resource {
     type       = "build"
@@ -129,6 +133,7 @@ resource "xray_watch" "all-projects" {
   name        = "projects-watch"
   description = "Watch all the projects"
   active      = true
+  project_key = "testproj"
 
   watch_resource {
     type       = "all-projects"
@@ -152,6 +157,7 @@ resource "xray_watch" "all-projects-with-filters" {
   name        = "projects-watch"
   description = "Watch all the projects with Ant patterns filter"
   active      = true
+  project_key = "testproj"
 
   watch_resource {
     type       = "all-projects"
@@ -180,6 +186,7 @@ resource "xray_watch" "project" {
   name        = "project-watch"
   description = "Watch selected projects"
   active      = true
+  project_key = "testproj"
 
   watch_resource {
     type = "project"